LoadScene improvement
The GameController's LoadScene method will now yield one frame to give time to a running transition to 'yield break' before calling Unload on the scene. This is useful to call LoadScene from the Load coroutine of a scene.

usage in the Load coroutine :
if (condition){
GameController.LoadScene("scenename");
yield break;
}